Use to_add for renames - fixup in ghost's BAIN refresh:
As far as the rest of the code is concerned those are to be added not
redrawn, although as seen redraw and to_add mostly come in pairs - very
few occurrences of either which is a good sign for RefrData.

`new_changed` to track redraw | to_add:

Once the dust settles a final test for RefrData would be to use to_add
and redraw separately in RUI (instead of dumping them together). We
would need to add code in __ior__ for that - and centralize RefrData
attributes access which are now few (41/50 occurrences in code/in total
of to_add/to_del/redraw).

Mopy/bash/bolt.py: I should be using abs_path here for ghosts - this
can explain some ghostly glitches in BAIN refresh - more on the latter
in next commit (ghosts is the single worst source of complications in
the codebase). From cd2460c199d20a0ff615d904c31ffcd2c9427117

Mopy/bash/basher/__init__.py: regression from
332e213498c377a61477adb04301ed81ed0603c5 - renames could be empty.
More work on this in next commit.

`dest_dir` renames - will come handy soon

Under #353