Use rename_operation for hiding:
The complication was that the destination directory may differ per mod -
ren_args should become a RenStruct as already noted - the rest is pretty
straight forward. move_to (to be retired) is replaced by shellMove - one
use of refresh down. Note:

- move_to -> Path.moveTo does a few things - I kept the makedirs call,
the rest should not be needed
- the explicit check for None was dropped in try_rename - we don't want
an empty filename here
- I pass `ren_parent=self` from try_rename - the only use that could
maybe need to pass the actual parent (the details panel - not the
uilist) is _rename_detail_item, but that one did not pass anything
anyway.