Drop BsaInfos.refresh override:
See comments in 9d61dbababefbc308de033aeba833715a5902786 - this is a
satisfactory solution as indicated by the fact that the ex link_frame
attributes became private and the signature of warning_args was
simplified. And of course by dropping a full override of refresh complete
with super and accessing the (ideally) private attributes of RefrData.
Finalizes bb78ffcb5cf34c0740aa3576a2bad088e2e7cc63 as as is now plain
those caches are not link_frame responsibilities - see also #542. Those
caches need more work (like invalidate for instance...) but that's #701.

Mopy/bash/bosh/bsa_files.py:
Probably needs to be an FName - for one for the 'Fallout - Misc.bsa'
comparison. It might be worth it to consolidate those caches a bit.