AFileInfo.fs_copy:
The only case `destDir == store_dir` is via copy_to in File_Duplicate
(and possibly in Mod_Snapshot.Execute that allows snapshots in the
store_dir). Initially I moved the handling of the ghosted path in the
call sites but then I bit the bullet and absorbed it in
get_rename_paths - I would anyway need the latter for the SaveInfo
override - so here we are - fs_copy became another thin wrapper around
get_rename_paths - apart from #241 noted issues - but this is nit for
now.