Bump dependencies (except PyInstaller :/)
A bunch of bugfixes, but some security fixes as well (e.g. lxml).

Bump dependencies

Some updates in here are needed for running on py3.14, so:

Under #684

Add a note regarding chardet

Sad situation all around :/ Let's avoid the controversial versions for
now, not like we're bottlenecked on chardet performance since our
performance optimizations that led to us mostly avoiding calling into it
in the first place. In the far future we may want to look into dropping
it entirely even, if only for performance.