Remove __setUI:
Note this drops get/set_item_data accessors as a bonus.

I added a UIList pointer to the list control as it was already in init
scope - this permitted me to drop the uilist param from insert_update_item.
So there is a chain UIList <- UIListCtrl <- ListItemFormat which I think
is the most elegant solution to this long standing de-wx comment.
Note that I removed the `if not allow_cols` check - should never be
empty (we could use an iterator here?). Not sure about the
`self._native_widget.GetTextColour()` - should his be cached or is
dynamical?

Under #190