Allow merging ESM-flagged plugins into the BP
These are important for FO3/FNV, where non-ESM plugins do not handle
navmeshes and facegen properly.

Most likely the reason we disallowed this initially was because these
are often masters of some other plugins (as the name ES_M_ indicates),
but WB has since learned to refuse merging of plugins that are masters
of non-mergeable plugins (since non-ESMs can be masters too, this was
required for correctness anyway). Since there are advantages to
allowing this and no disadvantages cited yet, let's try allowing this
and see if anything breaks.

Closes #727

Mopy/bash/brec/complex_subrecords.py:
FO4: Fix the BP breaking if OBTS contains FormIDs

Forgot to implement hasFids for this subrecord, so it was always
considered to contain no FormIDs, which meant it never got to mapFids,
which blew up once we got to the short-mapping phase since it also never
got to contribute its newly introduced FormID to the BP's master table.

Mopy/bash/bosh/__init__.py:
Avoid modifying the source BSA for BSA redirection

On global installations (like the AUR package and our future deployment
on Windows too), the source BSA won't be writable, so this will just
fail. Set the timestamp on the target instead.