TES5/SSE: Add Remove Load Screen Models tweak
Tested in-game & the model must be set to a null form ID for it to work
as intended; the game seems to grab the model from elsewhere otherwise
(didn't test further, but I'd guess it's from the last plugin to define
the model for the record). The other attributes are removed when setting
the model to NONE in the CK, so remove them as well.

Part of #602.