Move Store to basher:
After removing the local bosh import from balt I realized that
`.unique_store_key` was really accessed in two places - and that every
time the data store itself was in scope. This meant I don't need Store
anymore. Historically Store was introduced to communicate information
on refresh #353, and to store info on the new tabs - see
46aa64fa7847b1884d372c172d659cbf91b33571. The use in refresh was
a refactoring artifact needed to trace and eventually prune the places
where a refresh_info=True was passed. Since then all these calls were
centralized in propagate refresh and we are able to drop the ubiquitous
Store. A bosh.Store enum whose members are the data store singletons
(possible with Enums?) is what we would really like to have for #600
but that's just cosmetics for now (at least before #190 is done). The
weird `store_key = self` assignments are for minimizing the conflicts
with a future commit as this came up while rebasing.