WIP absorb tabInfo in _BashTab:
It's not (ever) an accident that the code looked ugly (as in u'',
camelCase and worse - globals(), things like *tabInfo[k][:2] etc):

https://www.joelonsoftware.com/2005/05/11/making-wrong-code-look-wrong/

Well it was not wrong per se - but having two data structures is bad
enough. Now that _BashTab is in basher we can unify those. WIP as we
can do some more - in particular which is the one right way to get the
current tab? Going via the title seems a bit odd.