Fix for record-children signatures being ignored:
Partially reverts 3badb8fd7e161445cd8bb8423a5f3dc3165aa679.
A better solution could be to modify merge_records, which knows which
records were actually merged, to return a list of signatures to include
in load/read factories for subsequent mods. This could be done via a
_MergeFact(LoadFactory) that exports a merged_sigs attribute, but it’s
not immediately clear how to update the merge_records overrides to keep
this set in sync.

Huge thanks to @sibir-ine and @slowpard for bisecting and code.

Closes #687.