Move pyfiglet header generation to gen_ini
bash_default.ini no longer needs to be updated externally on version
bump, the version header can be generated along with it. This makes
pyfiglet a runtime requirement, but we should make it optional since
it's not absolutely necessary. Use it to generate the section headers as
well & move inside sections so they properly collapse in, e.g.,
Notepad++.

Under #657.