[!]Call RUI on the detail item in case of modifying tags:
It's still difficult to follow where tags are used in refresh (so if we
need to call refresh) but IIUC we only need to call RUI - [!]. We do
need to call the latter however, and no need to call it OnDismiss (less
event handling at that) just after the action is executed.

Note I removed the refresh call in Tags_CopyToBashTags.Execute - seems
unneeded still [!] - also cause it's a last minute addition to this
branch but I need it for completeness - tag refresh should be bug free
now.