[!]Move loading to DataStore.refresh:
The difficult part to check is if the super.refresh() and
_refresh_from_data_dir() commute - I believe yes - they should anyway
(load - that is now moved in super - should not depend on the latter
and _refresh_from_data_dir should not need the installers - still [!]).
I moved last marker creation to take place after load (so we load the
marker from the dat if it's not first load ever) - used in refreshOrder.
I finally addressed the pre-historic† FIXME with a draft of docs (status
handling will be more centralized as part of #701).

Inline TableFileInfos - it's _AFileInfos vs IData now, table remains
are all but completely absorbed in the former‡ (flat is better than all
the nested init and refresh calls).

†: a4434cc2bf149b913cbeed3844208a6d026cba91
‡: there is also _TabledInfo and RefrIn.from_tabled_infos, and that
should be it

Closes #716