Inline _list_store_dir and skip stat'ing loaded installers:
A final method with one use - hmmm. We should not need to stat
installers for which we have an info on boot - we already stated those
in __setstate. More profiling is needed to really optimize (optimize
irefresh deserves a small issue).

Common code in DataStore.__init__:

In particular centralize set_store_dir.