Encapsulate BOM handling in Path:
That one was the last big remain of #232 - unicode handling in Bash is
not an issue anymore as it was in python2. What remains is plugins/saves
string encoding and remove Paths from pickle files (and open_bom('w')
calls) - #734.

bolt:
Path (#368) is not an issue anymore either - the big problem was Path
being used as the data stores key - and this was a big issue (in
particular #543). We can lazily uproot some Paths here and there and
that's it for now -ah yes and move it to a top level module. Some little
(style) optimization in forward_compat_path_to_fn - that `lambda x: x` -
duh. Plus prepare it for when we drop pickling FName as Path.

Closes #232 and under #734 - collected the todos for the latter (WIP,
some backwards compat can stay for ever like tags - some other must go).