[!]rightFileType: use an extension check rather than regex - fixups:
This isn't really faster, but it's cleaner. For one, it made me realize
that first backups (.esf files) were added to the save list if present
in the saves dir - certainly not intended. But this was on my mind for
a long time - Bash is using basically extension check but rightFileType
hides that (regexes are opaque for mere mortals) - I wanted this to be
more explicit as the extensions per Info are also used in many other
places (name validation, creating files etc etc) and need centralizing.
Another reason to make extension handling explicit is #701 - a super
store should keep dicts mapping extension of sub stores to store
processors. Don't mind ModInfos clunky implementation, this is still
WIP - we set the stage here for eliminating almost all remaining
NotImplemented DataStore APIs (and for a good cleanup of Infos extension
handling).

Nit in valid_save_exts:

Or rather drop valid_save_exts - see above

Fixup for saves first backups appearing in saves list:

To fix I passed allow_ext to rightFileType - I could copy the
implementation passing another set but I prefer keeping it in one place
- note that now parse_save_path in cosaves will also return None for
the cls._ess_skips which makes sense I think - still [!].

Note in Installers we fixed a bug - we were adding the inode not the
filename to the list of omods.

Also note in ScreenInfos override I FName the filename making the
__contains__ check case insensitive - counts as another fixup.