Compute _valid_exts_re for the infos from file_exts:
Deduplicate the logic - could do some more, but that requires generic
types. I initially used the _store() accessor on the infos, but stores
are None during tests, so that blew up. I therefore went all the way to
the correct solution, that is move file_exts into the infos, but that
needed a bit more work. Note we also avoid the _store coupling, which
highlights a dependency of infos on their store that should stay as
minimal as possible. Note that DefaultIniInfo does not define file_exts
- that's good (remember: no rename in the INI tab). One motivation here
was reducing occurrences of _valid_exts_re itself, since rename-input
validation still needs to be finalized. Another was reducing
occurrences of bush.game extension attributes - note this also absorbed
unhide_wildcard (which would sometimes add an extra '*').

Under #336