Encapsulate unique_key/validate_name WIP:
check_store's unique use was in FileDuplicate - took the logic out
of validate_name in Execute and this got me rid of the AFileInfo
override - nice. I also moved the makedirs call right in set_store_dir -
it certainly does not belong to a Link.Execute but see #478 / #417
Not sure if we should allow overwriting infos in store_dir in
FileDuplicate, but then let's keep it as it is.