WIP unique_key:
Based on the observation that what we need is the original extension,
let's add this logic to unique_key - this drops the ext parameter and
the force_ext param of try_rename. We now also pass the complete FName
to try_rename -> unique_key instead of the root - that proved difficult
to uproot, as it belongs to the (missing) RenStruct (it seems the first
few init arguments or fields could be (inf, newfn, ...)). Passing the
filename reads more naturally (noise like `r,e = ...` is gone), but we
needed to recast most of the logic for extension checking. The lowest
hanging fruit was the Installer override of validate_filename_str -
IIUC, a check in super will do fine[*]. But then there are default inis;
seeing that uses of file_exts are in validation, I added it to
DefaultIniInfo and moved the handling to base, which eliminated
_valid_exts_re.

[*]: Turns out the tests were failing, so here is a second round of
cleanup. use_default_ext is handed back to the UI, while extension
handling is moved to ListInfo.validate_filename_str. The latter becomes
stricter by only allowing file_exts subsets for ListInfo subclasses,
while on the other hand ignoring allowed_exts if the class has no
file_exts. This is final (yes), as it eliminated the BAIN override,
complete with all its hacks (can you spot 3?).

Mopy/bash/basher/installer_links.py: since allowed_exts are now ignored
for projects, we can pass the (random) extension here to force keeping
the extension for archives. As unique_key now also does the splitting,
we no longer need the is_archive hack here.

Note: no_dir=False was only passed in combination with
disallow_override=True. is_dir implies exists, so it will fail with a
better message - this dropped no_dir.

Mopy/bash/basher/__init__.py: the only use of **val_kwargs was in
InstallersList, but now that we enforce file_exts and those are the
read extensions, we no longer need to pass them. We do, however, need
to force the extension in _info_to_name.